refactor: rename definitions.go to manifest.go.
